Try to fix flickering time left annotation
This should fix (as in I'm not certain because there isn't any reliable way to reproduce this issue) the problem. I believe it was because `timers` map operated on a lesson's position in the view, which could change (although shouldn't unless there was a reload). Now it operates on the lessons id thus making sure there are no two timers operating on the same lesson. If this problem were to appear again it should only manifest as two separate lessons with timers, not a single one with a flickering timer.